requirements.txt
appdirs==1.4.4
attrs==21.2.0
berserk==0.10.0
black==19.10b0
certifi==2021.5.30
charset-normalizer==2.0.3
chess==1.6.1
click==8.0.1
Deprecated==1.2.12
idna==3.2
importlib-metadata==3.10.0
isort==5.9.2
mkl-fft==1.3.0
mkl-random==1.2.2
mkl-service==2.4.0
mypy-extensions==0.4.3
ndjson==0.3.1
numpy==1.20.3
pathspec==0.7.0
pip==21.1.3
python-dotenv==0.19.0
regex==2021.7.6
requests==2.26.0
setuptools==52.0.0.post20210125
six==1.16.0
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.10.0.0
urllib3==1.26.6
wheel==0.36.2
wrapt==1.12.1
zipp==3.5.0
